Fix type of min_value and max_value on DecimalField
These should at the very least allow Decimals. Technically you can send
in anything that's comparable to a Decimal, but I'm not sure if it makes
sense to allow floats. Could allow both ints and Decimals I guess?

+++ django-stubs/forms/fields.pyi
@@ -125,8 +125,8 @@ class DecimalField(IntegerField):
     def __init__(
         self,
         *,
-        max_value: Optional[int] = ...,
-        min_value: Optional[int] = ...,
+        max_value: Optional[Decimal] = ...,
+        min_value: Optional[Decimal] = ...,
         max_digits: Optional[int] = ...,
         decimal_places: Optional[int] = ...,
         required: bool = ...,
